Add suport for Dashboards.setOptions()
