+++ ts/masters-dashboards/dashboards.src.ts
@@ -50,6 +50,7 @@ import DataCursor from '../Data/DataCursor.js';
 import DataConverter from '../Data/Converters/DataConverter.js';
 import DataModifier from '../Data/Modifiers/DataModifier.js';
 import DataTable from '../Data/DataTable.js';
+import Defaults from '../Dashboards/Defaults.js';
 import Globals from '../Dashboards/Globals.js';
 import GridPlugin from '../Dashboards/Plugins/GridPlugin.js';
 import HighchartsPlugin from '../Dashboards/Plugins/HighchartsPlugin.js';
@@ -74,6 +75,7 @@ declare global {
         error: typeof Utilities.error;
         merge: typeof Utilities.merge;
         removeEvent: typeof Utilities.removeEvent;
+        setOptions: typeof Defaults.setOptions;
         uniqueKey: typeof Utilities.uniqueKey;
         version: typeof Globals.version;
         win: typeof Globals.win;
@@ -87,6 +89,7 @@ declare global {
         DataModifier: typeof DataModifier;
         DataPool: typeof DataPool;
         DataTable: typeof DataTable;
+        defaultOptions: typeof Defaults.defaultOptions;
         GridPlugin: typeof GridPlugin;
         HighchartsPlugin: typeof HighchartsPlugin;
         PluginHandler: typeof PluginHandler;
@@ -115,6 +118,7 @@ G.addEvent = Utilities.addEvent;
 G.error = Utilities.error;
 G.merge = Utilities.merge;
 G.removeEvent = Utilities.removeEvent;
+G.setOptions = Defaults.setOptions;
 G.uniqueKey = Utilities.uniqueKey;
 G.AST = AST;
 G.Board = Board;
@@ -126,6 +130,7 @@ G.DataCursor = DataCursor;
 G.DataModifier = DataModifier;
 G.DataPool = DataPool;
 G.DataTable = DataTable;
+G.defaultOptions = Defaults.defaultOptions;
 G.GridPlugin = GridPlugin;
 G.HighchartsPlugin = HighchartsPlugin;
 G.PluginHandler = PluginHandler;
